+++ packages/tendermint-rpc/src/tendermint34/responses.ts
@@ -8,6 +8,7 @@ export type Response =
   | AbciQueryResponse
   | BlockResponse
   | BlockResultsResponse
+  | BlockSearchResponse
   | BlockchainResponse
   | BroadcastTxAsyncResponse
   | BroadcastTxSyncResponse
@@ -70,8 +71,12 @@ export interface BlockchainResponse {
   readonly blockMetas: readonly BlockMeta[];
 }
 
-/** No data in here because RPC method BroadcastTxAsync "returns right away, with no response" */
-export interface BroadcastTxAsyncResponse {}
+/**
+ * No transaction data in here because RPC method BroadcastTxAsync "returns right away, with no response"
+ */
+export interface BroadcastTxAsyncResponse {
+  readonly hash: Uint8Array;
+}
 
 export interface BroadcastTxSyncResponse extends TxData {
   readonly hash: Uint8Array;